+++ FishMMO-Database/FishMMO-DB/Npgsql/Entities/Scene/Character/CharacterItemCooldownEntity.cs
@@ -9,7 +9,7 @@ namespace FishMMO.Database.Npgsql.Entities
 		public long CharacterID { get; set; }
 		public CharacterEntity Character { get; set; }
 		public string Category { get; set; }
-		public float CooldownEnd { get; set; }
+		public double CooldownEnd { get; set; }
 		public DateTime TimeCreated { get; set; }
 		public bool Deleted { get; set; }
 		public DateTime? TimeDeleted { get; set; }
